feat: Update npm dependencies
The code changes in the `package.json` file update the npm dependencies. This includes adding new dependencies like `@eslint/config-array`, `@eslint/object-schema`, and `glob`, as well as removing the dependency `@next-auth/prisma-adapter`. Updating the dependencies ensures that the project is using the latest stable versions and includes necessary packages for linting and file searching.